refactor: move domain to the `Domain` tree to follow Monarch's structure more closely
