+++ package.yaml
@@ -8,7 +8,6 @@ copyright:           "2024 Author name here"
 
 extra-source-files:
 - README.md
-- CHANGELOG.md
 
 # Metadata used when publishing your package
 # synopsis:            Short description of your package
