+++ src/Analysis/WebAssembly/Monad.hs
@@ -1,0 +1,2 @@
+module Analysis.WebAssembly.Monad where
+
