+++ src/Domain/WebAssembly.hs
@@ -1,0 +1,7 @@
+module Domain.WebAssembly(
+    module Domain.WebAssembly.Class,
+    module Domain.WebAssembly.ConstantPropagation
+  ) where
+
+import Domain.WebAssembly.Class
+import Domain.WebAssembly.ConstantPropagation
